backend/requirements.txt
fastapi[standard]
uvicorn
psycopg[binary]
slowapi
python-dotenv
python-jose
fastapi-mail
google-genai
python-multipart